move load_or_update_date and should_update
